chore(deploy) Fixes to deploy scripts
