+++ scripts/depcheck.js
@@ -15,15 +15,15 @@
 
 'use strict';
 
+const colors = require('colors');
 const fs = require('fs');
 const path = require('path');
-const semver = require('semver')
+
+const packages = {};
 
 const npmDirectory = path.resolve('.');
-const npmConfigFile = path.resolve(npmDirectory, 'package.json');
-const npmConfig = require(npmConfigFile);
-npmConfig.version.replace(/-.*/, '');
-const targetVersion = semver.clean(process.argv[2]);
-npmConfig.version = targetVersion;
-fs.writeFileSync(npmConfigFile, JSON.stringify(npmConfig, null, 2), 'utf8');
+
+const masterPackageFile = path.resolve(npmDirectory, 'package.json');
+const masterPackage = require(masterPackageFile);
+packages['package.json'] = masterPackage;
